Done all

+++ src/pages/DiscountCodeList.jsx
@@ -2,6 +2,8 @@ import { useEffect } from "react";
 import { useState } from "react"
 import { deleteDiscountCode, getAllDiscount } from "../services/DiscountService";
 import { Link } from "react-router-dom/cjs/react-router-dom";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faPlus, faTrash, faWrench } from "@fortawesome/free-solid-svg-icons";
 
 
 
@@ -14,7 +16,7 @@ const DiscountCodeList = () => {
     useEffect(
         () => {
             getAllDiscount().then(
-                response =>{
+                response => {
                     setVouchers(response.data)
                 }
             )
@@ -24,15 +26,17 @@ const DiscountCodeList = () => {
     // ---------------------------------------------------------
 
     const updateVoucher = (id) => {
-        
+
     }
 
     const deleteVoucher = (id) => {
-        deleteDiscountCode(id).then(
-            response => {
-                setR(Math.random())
-            }
-        )
+        if (window.confirm("Are you sure to delete this Discount code?")) {
+            deleteDiscountCode(id).then(
+                response => {
+                    setR(Math.random())
+                }
+            )
+        }
     }
 
     const addVoucher = () => {
@@ -42,43 +46,43 @@ const DiscountCodeList = () => {
     return (
         <div className="body-wrapper">
             <h2 className='text-center'>Voucher list</h2>
-                    <div className='container'>                    
-                    <Link to="/add-discount"><button className='btn btn-primary' style={{marginBottom : "10px"}}>Add voucher</button></Link>
-                        <table className='table table-striped table-bordered'>
-                            <thead>
-                                <tr>
-                                    <th className='col-md-2'>Code</th>
-                                    <th className='col-md-2'>Discount %</th>
-                                    <th className='col-md-2'>Available</th>
-                                    <th className='col-md-2'>Available</th>
-                                    <th className='col-md-2'>Status</th>
-                                </tr>
-                            </thead>
+            <div className='container'>
+                <Link to="/add-discount"><button className='btn btn-primary' style={{ marginBottom: "10px" }}><FontAwesomeIcon icon={faPlus} /></button></Link>
+                <table className='table table-striped table-bordered'>
+                    <thead>
+                        <tr>
+                            <th className='col-md-2'>Code</th>
+                            <th className='col-md-2'>Discount %</th>
+                            <th className='col-md-2'>Available</th>
+                            <th className='col-md-2'>Available</th>
+                            <th className='col-md-2'>Status</th>
+                        </tr>
+                    </thead>
 
-                            <tbody className="">
-                                {vouchers.map((voucher) => (
-                                    <tr key={voucher.id}>
-                                        <td className="text-danger">{voucher.code}</td>
-                                        <td className="text-danger">{voucher.discountPercentage}</td>
-                                        <td className="text-danger">{voucher.timesUsable}</td>
-                                        <td className="text-danger">{voucher.expiryDate}</td>
+                    <tbody className="">
+                        {vouchers.map((voucher) => (
+                            <tr key={voucher.id}>
+                                <td className="text-danger">{voucher.code}</td>
+                                <td className="text-danger">{voucher.discountPercentage}</td>
+                                <td className="text-danger">{voucher.timesUsable}</td>
+                                <td className="text-danger">{voucher.expiryDate}</td>
 
-                                        <td>
-                                            <button onClick={() => updateVoucher(voucher.id)} className='btn btn-info'>
-                                                Update
-                                            </button>
-                                            <button onClick={() => deleteVoucher(voucher.id)} style={{ marginLeft: '10px' }} className='btn btn-warning'>
-                                                Delete
-                                            </button>
-                                            <button className='btn btn-success' style={{ marginLeft: '10px' }}>
-                                                Details
-                                            </button>
-                                        </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
-                    </div>
+                                <td>
+                                    <Link to={`/update-discount/${voucher.id}`}>
+                                        <button className='btn'>
+                                            <FontAwesomeIcon icon={faWrench} />
+                                        </button>
+                                    </Link>
+
+                                    <button onClick={() => deleteVoucher(voucher.id)} style={{ marginLeft: '10px' }} className='btn'>
+                                        <FontAwesomeIcon icon={faTrash} />
+                                    </button>
+                                </td>
+                            </tr>
+                        ))}
+                    </tbody>
+                </table>
+            </div>
         </div>
     )
 }
